Create validations, biews for new controllers and simple filter for rooms

+++ db/migrate/20190628214824_create_hotels.rb
@@ -2,7 +2,7 @@ class CreateHotels < ActiveRecord::Migration[5.2]
   def change
     create_table :hotels do |t|
       t.string :name
-      t.decimal :stars
+      t.integer :stars
       t.text :description
       t.references :city, foreign_key: true
